Simplify bindings
This fixes a problem which caused YNAB transactions to be identified as orphans
when the id of a N26 transaction would change. The old mapping logic did not
work as expected because the VisibleTs of the 'new' N26 transaction was actually
smaller than the 'old' one. This commit removes the grouping logic altogether
thus making sure this cannot happen again.

This also simplifies the way bindings are persisted -> only bindings of active
N26 transactions are now saved (instead of all of them). This reduces the number
of bindings and hopefully removes other potentially nasty problems.